+++ 0x08-recursion/main.h
@@ -15,6 +15,6 @@ int _pow_recursion(int x, int y);
 
 int _sqrt_recursion(int n);
 
-int find_sqrt(int n, int st, int end);
+int find_sqrt(long n, long st, long end);
 
 #endif /* MAIN_H */
